BUG: use int instead of float for DICOM writing
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=19338
git-svn-id: http://svn.slicer.org/Slicer4/trunk@19338 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Modules/CLI/CreateDICOMSeries/CreateDICOMSeries.cxx
@@ -32,8 +32,8 @@ int main( int argc, char* argv[] )
 {
   PARSE_ARGS;
 
-  typedef itk::Image<float, 3>                                 Image3DType;//changed from short to float for bright MRI images
-  typedef itk::Image<float, 2>                                 Image2DType;
+  typedef itk::Image<int, 3>                                   Image3DType;//changed from short to int for bright MRI images
+  typedef itk::Image<int, 2>                                   Image2DType;
   typedef itk::ImageFileReader<Image3DType>                    ReaderType;
   typedef itk::ShiftScaleImageFilter<Image3DType, Image3DType> ShiftScaleType;
   typedef itk::ImageFileWriter<Image2DType>                    WriterType;
